+++ 526-Beautiful-Arrangement/main.cpp
@@ -1,66 +1,89 @@
 #include <iostream>
-#include <stack>
-#include <tuple>
-#include <vector>
+#include <queue>
 
 using namespace std;
 
 class Solution {
 public:
+    // v1: lower space, but slower?
     static int countArrangement(int N) {
-        int num_perfect{};
-
-        stack<tuple<int, int, int>> num_level_idx;
+        deque<int> nums_left;
         for (int i = 1; i <= N; ++i) {
-            num_level_idx.emplace(i, 1, i);
+            nums_left.emplace_back(i);
         }
 
-        vector<bool> in_use(N + 1, false);
-        vector<int> prev(N + 1, 0), prev_idx(N + 1, 0);
-        while (!num_level_idx.empty()) {
-            auto num = std::get<0>(num_level_idx.top());
-            auto level = std::get<1>(num_level_idx.top());
-            auto idx = std::get<2>(num_level_idx.top());
-            num_level_idx.pop();
+        int num_perfect{};
+        nextNum(num_perfect, N, 1, nums_left);
 
-            in_use[num] = true;
-            if ((0 == num % level) || (0 == level % num)) {
-                if (N == level) {
-                    ++num_perfect;
-                    in_use[num] = false;
+        return num_perfect;
+    }
 
-                    while (1 == idx) {
-                        prev[num] = false;
-                        idx = prev_idx[idx];
-                    }
-                } else {
-                    int num_left = 0;
-                    for (int i = 1; i <= N; ++i) {
-                        if (in_use[i]) { continue; }
+    static void nextNum(int &num_perfect, const int N, const int level, deque<int> &nums_left) {
+        if (level == N) {
+            num_perfect += (((0 == N % nums_left.back()) || (0 == nums_left.back() % N)) ? 1 : 0);
+        }
 
-                        num_level_idx.emplace(i, level + 1, ++num_left);
+        // # of elements left to consider
+        int nums_size = nums_left.size();
 
-                    }
-                }
-            } else {
-                in_use[num] = false;
+        int tmp;
+        while (--nums_size >= 0) {
+            tmp = nums_left.back();
+            nums_left.pop_back();
 
-                while (1 == idx) {
-                    prev[num] = false;
-                    idx = prev_idx[idx];
-                }
+            if ((0 == tmp % level) || (0 == level % tmp)) {
+                // advance into next level
+                nextNum(num_perfect, N, level + 1, nums_left);
             }
+
+            // backtrack by adding back the removed element
+            nums_left.emplace_front(tmp);
+        }
+    }
+
+    // v0
+    /*
+    static int countArrangement(int N) {
+        deque<int> nums_left;
+        for (int i = 1; i <= N; ++i) {
+            nums_left.emplace_back(i);
+        }
+
+        return nextNum(N, 1, nums_left);
+    }
+
+    static int nextNum(const int N, const int level, deque<int> &nums_left) {
+        if (level == N) {
+            return ((0 == N % nums_left.back()) || (0 == nums_left.back() % N)) ? 1 : 0;
+        }
+
+        // # of perfect arrangements detected in this level
+        int num_perfect{};
+        // # of elements left to consider
+        int nums_size = nums_left.size();
+
+        int tmp;
+        while (--nums_size >= 0) {
+            tmp = nums_left.back();
+            nums_left.pop_back();
+
+            if ((0 == tmp % level) || (0 == level % tmp)) {
+                // advance into next level
+                num_perfect += nextNum(N, level + 1, nums_left);
+            }
+
+            // backtrack by adding back the removed element
+            nums_left.emplace_front(tmp);
         }
 
         return num_perfect;
     }
+     */
 };
 
 int main() {
-    const int N[]{2};
-
-    for (const auto n : N) {
-        cout << n << ": " << Solution::countArrangement(n) << endl;
+    for (int N = 1; N <= 15; ++N) {
+        cout << N << ": " << Solution::countArrangement(N) << endl;
     }
 
     return 0;